refactor: 이름변경 및 패키지 분리, 및 메시지 어트리뷰트 생성 메서드 분리
